requirements.txt
joblib
numpy
scikit-learn==1.6.1
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
mysqlclient
psycopg2-binary
python-decouple
gunicorn
flask-cors